Add test_basics_dtplyr. All tests passed.
Add a bind_cols for dtplyr tables b/c dplyr::bind_cols does not work with dtplyr tables, and the various joins do not work well without adding indices to the tables to join on.

Add dtplyr handling to ns, relu and intercept, b/c those are called in the basic tests.