Added test_matrix_dtplyr. All tests passed.
Added back casting the mode value to avoid coercing a factor column into a character column in the bake step for modeimpute. Added tests to collect dtplyr tables before any conversion for composition in the bake or juice step (otherwise, converting to matrix failed b/c it converted the dtplyr object and not the underlying table).